Rectif sur le controller de modification des sauces

+++ controllers/sauce.js
@@ -47,12 +47,14 @@ exports.deleteSauce = (req, res, next) => {
 exports.modifySauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
-            const filename = sauce.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, (error) => {
-                if (error) throw error;
-            })
+            if (req.file) {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, (error) => {
+                    if (error) throw error;
+                })
+            }
         })
-        .catch (error => res.status(500).json({ error }));
+        .catch(error => res.status(500).json({ error }));
     const sauceObject = req.file ? {
         ...JSON.parse(req.body.sauce),
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
